requirements.txt
alpaca-py>=0.22.0
openai>=1.50.0
pandas>=2.2.2
numpy>=1.26.4
scipy>=1.13.1
scikit-learn>=1.5.1
pydantic>=2.8.2
tenacity>=8.3.0
python-dateutil>=2.9.0.post0
tzdata>=2024.1
pytz>=2024.1
lxml>=4.9.4
Flask>=3.0.0
gevent>=24.2.1
gunicorn>=21.2.0
